% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{accept_qualification_request}
\alias{accept_qualification_request}
\title{The \code{AcceptQualificationRequest} operation approves a Worker's request for a Qualification}
\usage{
accept_qualification_request(QualificationRequestId, IntegerValue = NULL)
}
\arguments{
\item{QualificationRequestId}{The ID of the Qualification request, as returned by the \code{GetQualificationRequests} operation.}

\item{IntegerValue}{The value of the Qualification. You can omit this value if you are using the presence or absence of the Qualification as the basis for a HIT requirement.}
}
\description{
The \code{AcceptQualificationRequest} operation approves a Worker's request for a Qualification.
}
\details{
Only the owner of the Qualification type can grant a Qualification request for that type.

A successful request for the \code{AcceptQualificationRequest} operation returns with no errors and an empty body.
}
\examples{

}
