% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{delete_worker_block}
\alias{delete_worker_block}
\title{The \code{DeleteWorkerBlock} operation allows you to reinstate a blocked Worker to work on your HITs}
\usage{
delete_worker_block(
  WorkerId = "string",
  Reason = "string"
)
}
\arguments{
\item{WorkerId}{[required] The ID of the Worker to unblock.}

\item{Reason}{A message that explains the reason for unblocking the Worker. The Worker does not see this message.}
}
\description{
The \code{DeleteWorkerBlock} operation allows you to reinstate a blocked Worker to work on your HITs. This operation reverses the effects of the CreateWorkerBlock operation. You need the Worker ID to use this operation. If the Worker ID is missing or invalid, this operation fails and returns the message "WorkerId is invalid." If the specified Worker is not blocked, this operation returns successfully.
}
\examples{

}
