% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{disassociate_qualification_from_worker}
\alias{disassociate_qualification_from_worker}
\title{The \code{DisassociateQualificationFromWorker} revokes a previously granted Qualification from a user}
\usage{
disassociate_qualification_from_worker(WorkerId, QualificationTypeId,
  Reason = NULL)
}
\arguments{
\item{WorkerId}{The ID of the Worker who possesses the Qualification to be revoked.}

\item{QualificationTypeId}{The ID of the Qualification type of the Qualification to be revoked.}

\item{Reason}{A text message that explains why the Qualification was revoked. The user who had the Qualification sees this message.}
}
\description{
The \code{DisassociateQualificationFromWorker} revokes a previously granted Qualification from a user.
}
\details{
You can provide a text message explaining why the Qualification was revoked. The user who had the Qualification can see this message.
}
\examples{

}
