% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{get_file_upload_url}
\alias{get_file_upload_url}
\title{The GetFileUploadURL operation generates and returns a temporary URL}
\usage{
get_file_upload_url(AssignmentId, QuestionIdentifier)
}
\arguments{
\item{AssignmentId}{[required] The ID of the assignment that contains the question with a FileUploadAnswer.}

\item{QuestionIdentifier}{[required] The identifier of the question with a FileUploadAnswer, as specified in the QuestionForm of the HIT.}
}
\description{
The \code{GetFileUploadURL} operation generates and returns a temporary URL. You use the temporary URL to retrieve a file uploaded by a Worker as an answer to a FileUploadAnswer question for a HIT. The temporary URL is generated the instant the GetFileUploadURL operation is called, and is valid for 60 seconds. You can get a temporary file upload URL any time until the HIT is disposed. After the HIT is disposed, any uploaded files are deleted, and cannot be retrieved. Pending Deprecation on December 12, 2017. The Answer Specification structure will no longer support the \code{FileUploadAnswer} element to be used for the QuestionForm data structure. Instead, we recommend that Requesters who want to create HITs asking Workers to upload files to use Amazon S3.
}
\section{Accepted Parameters}{
\preformatted{get_file_upload_url(
  AssignmentId = "string",
  QuestionIdentifier = "string"
)
}
}

