% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{get_hit}
\alias{get_hit}
\title{The \code{GetHIT} operation retrieves the details of the specified HIT}
\usage{
get_hit(HITId)
}
\arguments{
\item{HITId}{The ID of the HIT to be retrieved.}
}
\description{
The \code{GetHIT} operation retrieves the details of the specified HIT.
}
\examples{

}
