% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{get_qualification_score}
\alias{get_qualification_score}
\title{The \code{GetQualificationScore} operation returns the value of a Worker's Qualification for a given Qualification type}
\usage{
get_qualification_score(QualificationTypeId, WorkerId)
}
\arguments{
\item{QualificationTypeId}{The ID of the QualificationType.}

\item{WorkerId}{The ID of the Worker whose Qualification is being updated.}
}
\description{
The \code{GetQualificationScore} operation returns the value of a Worker's Qualification for a given Qualification type.
}
\details{
To get a Worker's Qualification, you must know the Worker's ID. The Worker's ID is included in the assignment data returned by the \code{ListAssignmentsForHIT} operation.

Only the owner of a Qualification type can query the value of a Worker's Qualification of that type.
}
\examples{

}
