% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{get_qualification_type}
\alias{get_qualification_type}
\title{The GetQualificationTypeoperation retrieves information about a Qualification type using its ID}
\usage{
get_qualification_type(QualificationTypeId)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the QualificationType.}
}
\description{
The \code{GetQualificationType}operation retrieves information about a Qualification type using its ID.
}
\section{Accepted Parameters}{
\preformatted{get_qualification_type(
  QualificationTypeId = "string"
)
}
}

