% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{list_bonus_payments}
\alias{list_bonus_payments}
\title{The \code{ListBonusPayments} operation retrieves the amounts of bonuses you have paid to Workers for a given HIT or assignment}
\usage{
list_bonus_payments(
  HITId = "string",
  AssignmentId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{HITId}{The ID of the HIT associated with the bonus payments to retrieve. If not specified, all bonus payments for all assignments for the given HIT are returned. Either the HITId parameter or the AssignmentId parameter must be specified}

\item{AssignmentId}{The ID of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified}

\item{NextToken}{Pagination token}

\item{MaxResults}{}
}
\description{
The \code{ListBonusPayments} operation retrieves the amounts of bonuses you have paid to Workers for a given HIT or assignment.
}
\examples{

}
