% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{list_hi_ts_for_qualification_type}
\alias{list_hi_ts_for_qualification_type}
\title{The ListHITsForQualificationType operation returns the HITs that use the given Qualification type for a Qualification requirement}
\usage{
list_hi_ts_for_qualification_type(QualificationTypeId, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the Qualification type to use when querying HITs.}

\item{NextToken}{Pagination Token}

\item{MaxResults}{Limit the number of results returned.}
}
\description{
The \code{ListHITsForQualificationType} operation returns the HITs that use the given Qualification type for a Qualification requirement. The operation returns HITs of any status, except for HITs that have been deleted with the \code{DeleteHIT} operation or that have been auto-deleted.
}
\section{Accepted Parameters}{
\preformatted{list_hi_ts_for_qualification_type(
  QualificationTypeId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

