% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{list_qualification_requests}
\alias{list_qualification_requests}
\title{The ListQualificationRequests operation retrieves requests for Qualifications of a particular Qualification type}
\usage{
list_qualification_requests(QualificationTypeId = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{QualificationTypeId}{The ID of the QualificationType.}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
The \code{ListQualificationRequests} operation retrieves requests for Qualifications of a particular Qualification type. The owner of the Qualification type calls this operation to poll for pending requests, and accepts them using the AcceptQualification operation.
}
\section{Accepted Parameters}{
\preformatted{list_qualification_requests(
  QualificationTypeId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

