% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{list_worker_blocks}
\alias{list_worker_blocks}
\title{The \code{ListWorkersBlocks} operation retrieves a list of Workers who are blocked from working on your HITs}
\usage{
list_worker_blocks(
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{NextToken}{Pagination token}

\item{MaxResults}{}
}
\description{
The \code{ListWorkersBlocks} operation retrieves a list of Workers who are blocked from working on your HITs.
}
\examples{

}
