% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{list_workers_with_qualification_type}
\alias{list_workers_with_qualification_type}
\title{The ListWorkersWithQualificationType operation returns all of the Workers that have been associated with a given Qualification type}
\usage{
list_workers_with_qualification_type(QualificationTypeId, Status = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the Qualification type of the Qualifications to return.}

\item{Status}{The status of the Qualifications to return. Can be \code{Granted | Revoked}.}

\item{NextToken}{Pagination Token}

\item{MaxResults}{Limit the number of results returned.}
}
\description{
The \code{ListWorkersWithQualificationType} operation returns all of the Workers that have been associated with a given Qualification type.
}
\section{Accepted Parameters}{
\preformatted{list_workers_with_qualification_type(
  QualificationTypeId = "string",
  Status = "Granted"|"Revoked",
  NextToken = "string",
  MaxResults = 123
)
}
}

