% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_package.R
\docType{package}
\name{paws.mturk-package}
\alias{paws.mturk}
\alias{paws.mturk-package}
\title{paws.mturk: Amazon Mechanical Turk}
\description{
Amazon Mechanical Turk API Reference
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
