% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{reject_qualification_request}
\alias{reject_qualification_request}
\title{The \code{RejectQualificationRequest} operation rejects a user's request for a Qualification}
\usage{
reject_qualification_request(
  QualificationRequestId = "string",
  Reason = "string"
)
}
\arguments{
\item{QualificationRequestId}{[required] The ID of the Qualification request, as returned by the \code{ListQualificationRequests} operation.}

\item{Reason}{A text message explaining why the request was rejected, to be shown to the Worker who made the request.}
}
\description{
The \code{RejectQualificationRequest} operation rejects a user's request for a Qualification.
}
\details{
You can provide a text message explaining why the request was rejected. The Worker who made the request can see this message.
}
\examples{

}
