% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mturk_operations.R
\name{get_assignment}
\alias{get_assignment}
\title{The \code{GetAssignment} operation retrieves the details of the specified Assignment}
\usage{
get_assignment(AssignmentId)
}
\arguments{
\item{AssignmentId}{The ID of the Assignment to be retrieved.}
}
\description{
The \code{GetAssignment} operation retrieves the details of the specified Assignment.
}
\examples{

}
